Week 2: Viewer Prototype
In viewer.py, I added some simple code to the function load_model in order to make it show a 3D model. It requires an absolute path at the moment.
IMPORTANT: One line, self.add_axes_widget() was not letting me test loading the model, so I commented it out for now.
The load_model function is called in the __init__ function.

In main_window.py, I added code that can hopefully help us with week 3. It is all in docstrings, so it should not effect anything.