Download and Thumbail
For the download button: I made it work and send the .obj file to the users downloads folder.

Thumbnail: not currently set up with the data manager (did not want to mess with that before it worked). It makes an offscreen plot and takes a screenshot of it to have a preview. Eventually can turn this into a gif, but I thought a png could be easier for now.
Made a new folder to make this work, where hopefully all of the current thumbnails could be loaded to just like the tags and such with datamanager.