Edited Download and Gallery Button
Previous issue: the download button would download directly into the project and not the downloads folder.
Solution: for windows, I found that if the file already exists, it will not go to the downloads folder. For my computer, it now works, and will add numbers to the end of the file so that there is no duplicates.

Gallery button: I added some code to link to the data manager. As we do not have a way to add new content, I am unsure where to continue and will leave it commented out for now.