Gallery change
Main change was the Gallery view:
Increased the size of the icons as well as make the gallery more of a grid view.
I also had to edit the thumbnail generation to make sure that all of the model is seen regardless of size.

Pytests:
Made two successful pytests and started a third

Data management addition:
Due one of a failed pytests, I accidently added two models that don't exist, so I made a function that hopefully works well that either gets rid of or ignores the models that have no filepath that exists (I intended it to delete them but I cannot tell if it was successful).