Part 1: Vector Database setup
I edited the client_data_manager to add a new function to concatenate filenames and tags using the meta.json file. I tried to make it so that it should be easy ish for us to use a different file in case we don't want meta.json.
Currently testing (the validation script part):
using the scene_brain.py as a TEMPORARY testing place. I am making a pytest for this validation section to try to move towards requirements as well, but need a normal .py file to test it easier