Part 2: Scene brain (IN PROGRESS)
Created three functions in the SceneBrain class

OPTIONAL stopwords function:
if we wanted to remove stopwords from the user input, this function fills a set from a stopwords list.

extract_keywords:
this function does basic string operations to get rid of whitespace and optionally removes stopwords

keywords_to_vector: (IN PROGRESS)
outputs the best matches. Assumes that the client data has been called before this as it uses the vector database.
It finds matches with a score higher than 0.5, and if not will return a placeholder.