Part 2: pytest
I made a pytest for part 2, inputting "I want a green forest" and trying to get tree as the best match.
By dong this, I hopefully bugtested for most of the issues that could be present. The pytest is successful on my end, and the scene_brain.py is working as it needs based on the project outline