Readme update
I consistently got a DLL error when trying to start the code. While it seemed to be just me, I updated the readme in case it is not.
To fix this, I added "import torch" to the top of the main.py file so that it is the first thing imported. For some reason, this matters.
It is NOT in the official code as it is an error that only I have been getting